+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.64.0] - 2026-02-22
+
+### Added
+
+- Support for projection expressions. The main exported type is `ProjectionExpressionParams`,
+which equates to a list of `RawPath`s. This means that client code can specify `literal`
+paths also in the list of projections.
+- The optional `projection` field on `GetItemInput`.
+
 ## [0.63.0] - 2026-02-21
 
 ### Added
@@ -814,6 +823,7 @@ intuitive.
 - Initial release of the package! Move the implementation work in progress from another
 project to here.
 
+[0.64.0]: https://github.com/infra-blocks/ts-aws-dynamodb/compare/v0.63.0...v0.64.0
 [0.63.0]: https://github.com/infra-blocks/ts-aws-dynamodb/compare/v0.62.0...v0.63.0
 [0.62.0]: https://github.com/infra-blocks/ts-aws-dynamodb/compare/v0.61.0...v0.62.0
 [0.61.0]: https://github.com/infra-blocks/ts-aws-dynamodb/compare/v0.60.0...v0.61.0